refactor(services): move slack
