+++ src/app/api/api_v1/endpoints/users.py
@@ -14,7 +14,7 @@ from app.models import Provider, User, UserScope
 from app.schemas.login import TokenPayload
 from app.schemas.users import Cred, CredHash, UserCreate
 from app.services.github import gh_client
-from app.services.slack import slack_client
+from app.services.notifications.slack import slack_client
 from app.services.telemetry import telemetry_client
 
 router = APIRouter()
